fields added in existing entities and new files deleted,linting done
